+++ src/facade/mod.rs
@@ -1,5 +1,5 @@
-pub mod list_handler;
-pub mod help_handler;
 pub mod add_handler;
-pub mod misc_handler;
 pub mod clear_handler;
+pub mod help_handler;
+pub mod list_handler;
+pub mod misc_handler;