+++ src/features/mod.rs
@@ -1,4 +1,4 @@
-pub mod list;
-pub mod misc;
 pub mod add;
 pub mod clear;
+pub mod list;
+pub mod misc;