+++ src/util/mod.rs
@@ -1,5 +1,5 @@
+pub mod config;
+pub mod display;
 pub mod enums;
 pub mod helpers;
 pub mod models;
-pub mod config;
-pub mod display;